Move the DummyClient to the verde/utils module (#190)
This was in verde/model_selection for some reason. It should be in utils
instead and not public.
