+++ verde/utils.py
@@ -15,6 +15,28 @@ except ImportError:
 from .base.utils import check_data, n_1d_arrays
 
 
+class DummyClient:  # pylint: disable=no-self-use,too-few-public-methods
+    """
+    Dummy client to mimic a dask.distributed.Client for immediate local
+    execution.
+
+    >>> client = DummyClient()
+    >>> client.submit(sum, (1, 2, 3))
+    6
+    >>> print(client.scatter("bla"))
+    bla
+
+    """
+
+    def submit(self, function, *args, **kwargs):
+        "Execute function with the given arguments and return its output"
+        return function(*args, **kwargs)
+
+    def scatter(self, value):
+        "Does nothing but return the input"
+        return value
+
+
 def parse_engine(engine):
     """
     Choose the best engine available and check if it's valid.
